requirements.txt
aiohttp==3.10.9
python-dotenv==1.0.1
aiofiles==24.1.0

